pyproject.toml
[tool.poetry]
name = "project-library"
version = "0.1.0"
description = ""
authors = ["kr.pratik@ymail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
